[dev]: Replace MainMessage stories with actual button
