+++ stories/ButtonContent/ContentWrapper.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+const ButtonContentWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 0.625rem;
+`
+export default ButtonContentWrapper
