+++ stories/ButtonContent/index.ts
@@ -1,0 +1,4 @@
+export { default as ButtonContentFrontendOriented } from './FrontendOriented';
+export { default as ButtonContentComponentDriven } from './ComponentDriven';
+export { default as ButtonContentArchitectureAware } from './ArchitectureAware';
+export { default as ButtonContentDesignAware } from './DesignAware';
